Linux test failure.

+++ tests/test_path.py
@@ -6,6 +6,7 @@ import stat
 import time
 
 import pytest
+import sys
 
 from dkfileutils import path
 from yamldirs import create_files
@@ -131,8 +132,11 @@ def test_chmod():
         root = path.Path(_root)
         (root / 'a').chmod(00400)  # only read for only current user
         # (root / 'a').chmod(stat.S_IREAD)
-        with pytest.raises(OSError):
-            (root / 'a').unlink()
+        if sys.platform == 'win32':
+            # doesn't appear to be any way for a user to create a file that he
+            # can't unlink on linux.
+            with pytest.raises(OSError):
+                (root / 'a').unlink()
         assert root.listdir() == ['a']
         (root / 'a').chmod(stat.S_IWRITE)
         (root / 'a').unlink()
